+++ example/filesystem/main.go
@@ -2,12 +2,11 @@ package main
 
 import (
 	"context"
-	"errors"
 	"flag"
 	"fmt"
-	"log"
+	"io"
 	"os"
-	"os/signal"
+	"time"
 
 	"github.com/MegaGrindStone/go-mcp/pkg/mcp"
 	"github.com/MegaGrindStone/go-mcp/pkg/servers/filesystem"
@@ -25,45 +24,33 @@ func main() {
 		os.Exit(1)
 	}
 
-	srv, err := filesystem.NewStdIOServer(*path)
+	srvReader, srvWriter := io.Pipe()
+	cliReader, cliWriter := io.Pipe()
+
+	cliIO := mcp.NewStdIO(cliReader, srvWriter)
+	srvIO := mcp.NewStdIO(srvReader, cliWriter)
+
+	go srvIO.Start()
+	go cliIO.Start()
+
+	srv, err := filesystem.NewServer(srvIO, *path)
 	if err != nil {
 		fmt.Println("Error: failed to create filesystem server:", err)
 		os.Exit(1)
 	}
 
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	c := newClient(ctx)
-	cli := mcp.NewStdIOClient(c, srv.MCPServer)
-
-	sigChan := make(chan os.Signal, 1)
-	signal.Notify(sigChan, os.Interrupt)
-	go func() {
-		<-sigChan
-		cancel()
-	}()
-
-	readyChan := make(chan struct{})
 	errsChan := make(chan error)
-	go func() {
-		err := cli.Run(ctx, c, c, readyChan, errsChan)
-		if err != nil {
-			if !errors.Is(err, context.Canceled) {
-				log.Print(err)
-			}
-		}
-	}()
-	go func() {
-		for err := range errsChan {
-			fmt.Print(err)
-		}
-	}()
+	srvCtx, srvCancel := context.WithCancel(context.Background())
 
-	fmt.Println("Initializing session...")
-	<-readyChan
-	fmt.Println("Session initialized!")
+	go mcp.Serve(srvCtx, srv, errsChan,
+		mcp.WithServerPingInterval(30*time.Second),
+		mcp.WithToolServer(srv),
+	)
 
-	c.run()
-	fmt.Println("Exiting...")
+	cli := newClient(cliIO)
+	go cli.run()
+
+	<-cli.done
+
+	srvCancel()
 }