+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enhanced `SSEServer` to implement the `ServerTransport` interface.
 - Updated `SSEClient` to implement the `ClientTransport` interface.
 - Everything server now utilizes the `Transport` interface.
+- Filesystem server now utilizes the `Transport` interface.
 
 ### Removed